add account, entries, transfer module

+++ database/migrations/000002_account_setup.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "accounts" (
     id BIGSERIAL PRIMARY KEY,
     user_id INTEGER NOT NULL,
-    balance MONEY NOT NULL DEFAULT 0.00,
+    balance DOUBLE PRECISION NOT NULL DEFAULT 0.00,
     currency VARCHAR(10) NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES users(id)
@@ -13,7 +13,7 @@ UNIQUE (user_id, currency);
 CREATE TABLE "entries" (
     id BIGSERIAL PRIMARY KEY,
     account_id INTEGER NOT NULL,
-    amount MONEY NOT NULL,
+    amount DOUBLE PRECISION NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (account_id) REFERENCES accounts(id)
 );
@@ -22,7 +22,7 @@ CREATE TABLE "transfers" (
     id BIGSERIAL PRIMARY KEY,
     from_account_id INTEGER NOT NULL,
     to_account_id INTEGER NOT NULL,
-    amount MONEY NOT NULL,
+    amount DOUBLE PRECISION NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (from_account_id) REFERENCES accounts(id),
     FOREIGN KEY (to_account_id) REFERENCES accounts(id)
